Added files via upload
Hey. Here's a new idea for the GUI. I studied the assignment for the past few days and this is what I've got for inputs. These components are as simple as the what we've learned in the class. I'm still working on putting it on code. If you have code functions, it would be great.